pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "conduit-naas"
version = "0.1.0"
description = "Self-serve private networking for any cloud SaaS"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [{name = "Sridhar Sikha", email = "sridhar.sikha@gmail.com"}]
keywords = ["privatelink", "vpc", "networking", "saas", "self-serve", "kafka", "snowflake"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Networking",
]

dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "sqlalchemy>=2.0",
    "asyncpg>=0.29",
    "redis>=5.0",
    "confluent-kafka>=2.3",
    "boto3>=1.34",
    "PyJWT[crypto]>=2.8",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "typer>=0.9",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio", "httpx", "ruff", "mypy"]

[project.scripts]
conduit = "conduit.cli:app"

[tool.ruff]
line-length = 120
